fix(audit): 多面的完全監査で発見した実害・堅牢性・整合の欠陥を一括根治
8観点×並行監査（並行性/金額・cron時刻/冪等・スロット/perf・客側UI/a11y・店舗UI/null安全・
DB整合・SEO・依存/env）で発見した真の欠陥を根治。セキュリティ実装順序・IDOR・署名検証・
env fail-secure・依存脆弱性(critical/high=0)は健全を確認。

🔴 金銭損失:
- ポイント利用予約のキャンセルで控除済みポイントが返還されず永久消失していた（顧客キャンセル
  /api/booking/[id]/cancel・管理者キャンセル /api/admin/booking-status の両経路）。CAS 成功後に
  service_role で返還行を挿入。cancel 側は所有権チェックで非null保証の user.id を用い到達不能分岐を排除。
- points_used に .int() を追加（小数ポイント混入で台帳破損を予防）。

🟡 堅牢性/整合:
- 売上分析(admin/analytics)が月次クエリ失敗を「売上¥0」に偽装表示 → 失敗時 throw して error.tsx へ。
- 施設公開トグル(admin/settings)が fetch throw でボタン永久固着 → try/catch/finally で固着解消＋失敗明示。
- 問診票(IntakeForm) checkbox 型の描画分岐欠落で必須時に送信不能・fields 非配列で白画面・label 未関連付け
  → checkbox 実装/Array.isArray 防御/aria 関連付け。WaitlistForm の label 関連付けも修正。
- recommendations の limit が NaN/負数で不正クエリ化(500) → 1〜12 正規化。
- facility_profiles の本番実在 8 列が types/snapshot に欠落（列レベル drift・cast で tsc 盲点化）→ 型/snapshot 同期。
- sitemap に noindex の /compare を掲載していた矛盾を除去。

検証: tsc=0 / 全jest 4854 pass・グローバル branch 100%(5729/5729) / schema-snapshot 同期・
契約ドリフト緑。本番SQL適用は不要（drift は repo 同期のみ・8列は本番既存）。

+++ src/lib/validations-booking.ts
@@ -52,7 +52,7 @@ export const bookingSchema = z.object({
   phone: z.string().regex(phoneRegex, '正しい電話番号を入力してください').or(z.literal('')).optional().nullable(),
   note: z.string().max(500, '備考は500文字以内で入力してください').optional(),
   total_price: z.number().min(0).max(9999999).nullable(),
-  points_used: z.number().min(0).max(9999999).optional(),
+  points_used: z.number().int().min(0).max(9999999).optional(),
 });
 
 export type BookingFormData = z.infer<typeof bookingSchema>;
